requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
google-generativeai
PyMuPDF
python-docx
pytesseract
celery
redis
Pillow

